add more css

+++ client/src/components/customerUserProfile/customerUserProfile.js
@@ -14,6 +14,7 @@ export default function CustomerUserProfile(props) {
     const [errors, seterrors] = useState([]);
     const [emailerror, setemailerror] = useState("");   
     const[loginMsg] = useState(localStorage.getItem("loginMsg"));//this sets logout msg if any
+    const[success,setsuccess] = useState("")
 
     window.onload = function() {
         localStorage.removeItem("loginMsg")
@@ -58,7 +59,7 @@ export default function CustomerUserProfile(props) {
             document.getElementById('cancelButton').setAttribute('hidden','true');
             document.getElementById('resetButton').removeAttribute('hidden');
             document.getElementById('editButton').removeAttribute('hidden');
-            localStorage.setItem('success', 'You have successfully updated your profile!');
+            setsuccess('You have successfully updated your profile!');
             props.history.push("/customer/profile");
         }
         if (response.data.errors) {
@@ -76,9 +77,11 @@ export default function CustomerUserProfile(props) {
         props.history.push("/customer/profile/password-reset");
     }
     const logoutHandler = () => {
-        removeUser();
-        localStorage.setItem('logoutMsg', 'You have successfully logged out!');
-        window.location = '/customer/login'
+        if(window.confirm("Do you want to logout?")){
+            removeUser();
+            localStorage.setItem('logoutMsg', 'You have successfully logged out!');
+            window.location = '/customer/login'
+        }
     }
     return (
         <div name="userDetails-form">
@@ -86,7 +89,7 @@ export default function CustomerUserProfile(props) {
             <button className="btn it19951386-green-btn">My Orders</button>
             <form className="form-group it19951386-myForm">
                 <h2>User Profile</h2>
-                <button className="btn-danger btn" id="it19951386-logout-btn" onClick={logoutHandler}>Log out</button>
+                <button className="btn-danger btn it19951386-red-btn" id="it19951386-logout-btn" onClick={logoutHandler}>Log out</button>
                 {errors ? errors.map((error) => {
                     return (
                         <div class="alert alert-danger" role="alert">
@@ -102,6 +105,10 @@ export default function CustomerUserProfile(props) {
                 <div class="alert alert-success" role="alert">
                 {loginMsg}
                 </div> : null}
+                {success ? 
+                <div class="alert alert-success" role="alert">
+                    {success}
+                </div>: null}
                 <label>First name</label>
                 <input className="form-control" type="text" onChange={(e) => setfirstName(e.target.value)} value={fname} disabled/><br/>  
 
@@ -119,9 +126,9 @@ export default function CustomerUserProfile(props) {
 
                 <div className="it19951386-btn-section">
                     <input id="editButton" type="button" value="Edit Details" onClick={personalInfoForm} className="it19951386-green-btn btn" style={{marginRight:'5px'}}/>
-                    <input id="resetButton" type="button" value="Reset Password" onClick={resetButton} className="btn-primary btn"/>
+                    <input id="resetButton" type="button" value="Reset Password" onClick={resetButton} className="it19951386-trans-green-btn btn"/>
                     <input id="updateButton" type="button" value="Update" onClick={editData} className="it19951386-green-btn btn" style={{marginRight:'5px'}} hidden/>
-                    <input id="cancelButton" type="button" value="Cancel" onClick={cancelUpdate} className="btn-danger btn" hidden/>
+                    <input id="cancelButton" type="button" value="Cancel" onClick={cancelUpdate} className="btn-danger btn it19951386-red-btn" hidden/>
                 </div>
             </form>
         </div>
